header: Rework the start and stop header functions
The current code assumes that we'll have only one RenderPicture call per
frame, and will emit the frame headers and footers each time we are called.

However, that's not true anymore with H264 since RenderPicture will be
called multiple times for a single frame.

Move the header print to BeginPicture and the footer to EndPicture to make
sure our output is properly formatted.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>