header: introduce a dumber helper to help with printing data
The current code uses the header_dump_format function to output each
variable of the MPEG2 libva structure.

However, that helper is not fit for H264, since we will have multiple level
of indentation and will have more complex structures to dump that cannot be
easily expressed.

Switch to a printf-like function that simply adds level of indentation
instead.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>