dump: Add H264 support
Add the support for the various bits and pieces for H264 support.

The structure is a bit weird, and unlike MPEG2, the frames won't be emitted
on stdout right when the libva call occurs, since we need both the Slice
and Picture Parameters buffers to accomodate the kernel wanting an index in
the DPB for the reference pictures, while libva uses the frame numbers.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>